requirements.txt
bcrypt==4.0.1
colorama==0.4.6
psycopg2-binary==2.9.6
python-dotenv==1.0.0
